refactor: move elo logic into ranked system controller
